+++ .gitignore
@@ -1,7 +1,8 @@
 env/*
 build/*
 dist/*
+__pycache__/*
 *.exe
-*.json
+settings.json
 *.zip
 test_path.py
